requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
Deprecated==1.2.14
idna==3.4
psutil==5.9.5
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.0.4
wrapt==1.15.0
